requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.2
daphne==4.2.0
Django==5.2.8
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.3
mysqlclient==2.2.7
paho-mqtt==2.1.0
passlib==1.7.4
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pymodbus==3.9.2
PyMySQL==1.1.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-snap7==2.0.2
requests==2.32.4
service-identity==24.2.0
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
zope.interface==7.2
